// src/enums/css.enum..ts
export enum ECursor {
  DEFAULT = "default",
  POINTER = "pointer",
  E_RESIZE = "e-resize",
  EW_RESIZE = "ew-resize",
  GRAB = "grab",
  GRABBING = "grabbing",
  HELP = "help",
  MOVE = "move",
  N_RESIZE = "n-resize",
  NE_RESIZE = "ne-resize",
  NESW_RESIZE = "nesw-resize",
  NS_RESIZE = "ns-resize",
  NW_RESIZE = "nw-resize",
  NWSE_RESIZE = "nwse-resize",
  NO_DROP = "no-drop",
  NOT_ALLOWED = "not-allowed",
  PROGRESS = "progress",
  ROW_RESIZE = "row-resize",
  S_RESIZE = "s-resize",
  SE_RESIZE = "se-resize",
  SW_RESIZE = "sw-resize",
  TEXT = "text",
  URL = "url",
  WAIT = "wait",
  ZOOM_IN = "zoom-in",
  ZOOM_OUT = "zoom-out",
}